no overwrite

+++ sphinx_gallery/gen_rst.py
@@ -877,7 +877,8 @@ def generate_file_rst(fname, target_dir, src_dir, gallery_conf,
             image_path_iterator = script_vars['image_path_iterator']
             stock_img = os.path.join(glr_path_static(), 'no_image.png')
             for _, path in zip(range(dummy_image), image_path_iterator):
-                copyfile(stock_img, path)
+                if not os.path.isfile(path):
+                    copyfile(stock_img, path)
 
     if gallery_conf['remove_config_comments']:
         script_blocks = [
